+++ src/lib.rs
@@ -1,5 +1,5 @@
-pub mod models;
 pub mod utils;
 pub mod handlers;
 pub mod routes;
 pub mod middlewares;
+pub mod models;
