+++ src/models/graphql.rs
@@ -1,0 +1,7 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize)]
+pub struct GraphQLRequest {
+    pub query: String,
+    pub operation_name: Option<String>,
+}
