+++ src/models/multipart.rs
@@ -1,0 +1,2 @@
+#[derive(Debug)]
+pub struct MultipartHandler;
